Improve helpers for error handling
This commit introduces a few optimizations and resources to make it
easier to define customized errors with debugging and user info.

This commit also includes a Dockerfile and a Makefile to build a
dockerized version of this lib. This way this is possible to use
golang workspaces along with docker-compose to locally develop
the whole system, including internal libs.