Make use of scopes mandatory for defining error fields
This change makes the use of errors.SystemScope and errors.UserScope
mandatory to propagate error fields. This decision is based on an attempt
to force developers to use the recommended standard without violating
the hierarchical structure of the errors.

This commit also introduces a new concept of Entry (Entries), since
we no longer support the same Opt interface for scoped-fields, it was
the natural replacement. The structure remains the same, the only change
is the removal of the `visibility` layer for Entries. It'll be controlled
only in the Scope level.